migrate types & MnemonicToAddress
