+++ local-interchain/interchain/types/chain.go
@@ -3,14 +3,7 @@ package types
 import (
 	"github.com/go-playground/validator"
 	"github.com/strangelove-ventures/interchaintest/v8/chain/cosmos"
-	"github.com/strangelove-ventures/interchaintest/v8/testutil"
 )
-
-// ConfigFileOverrides overrides app configuration files.
-type ConfigFileOverrides struct {
-	File  string        `json:"file" yaml:"file"`
-	Paths testutil.Toml `json:"paths" yaml:"paths"`
-}
 
 type Chain struct {
 	// ibc chain config (optional)
