+++ local-interchain/interchain/types/genesis.go
@@ -3,6 +3,12 @@ package types
 import (
 	"fmt"
 
+	"github.com/cosmos/cosmos-sdk/crypto/hd"
+	"github.com/cosmos/cosmos-sdk/crypto/keyring"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
+	"github.com/cosmos/cosmos-sdk/x/auth"
+	"github.com/cosmos/cosmos-sdk/x/auth/codec"
 	"github.com/cosmos/go-bip39"
 	"github.com/strangelove-ventures/interchaintest/v8/chain/cosmos"
 )
@@ -25,22 +31,50 @@ type Genesis struct {
 	StartupCommands []string `json:"startup_commands" yaml:"startup_commands"`
 }
 
-func GenerateRandomAccounts(bech32 string, num int) []GenesisAccount {
+func GenerateRandomAccounts(num int, bech32 string, coinType int) []GenesisAccount {
 	accounts := []GenesisAccount{}
 
 	for i := 0; i < num; i++ {
 		entropy, _ := bip39.NewEntropy(256)
 		mnemonic, _ := bip39.NewMnemonic(entropy)
 
-		// load mnemonic into cosmossdk and get the address
 		accounts = append(accounts, GenesisAccount{
 			Name:     fmt.Sprintf("account%d", i),
-			Amount:   "100000%DENOM%", // allow user to alter along with keyname?
-			Address:  "",              // TODO:
+			Amount:   "100000%DENOM%",
+			Address:  MnemonicToAddress(mnemonic, bech32, uint32(coinType)),
 			Mnemonic: mnemonic,
 		})
 
 	}
 
 	return accounts
+}
+
+func MnemonicToAddress(mnemonic, bech32 string, coinType uint32) string {
+	e := moduletestutil.MakeTestEncodingConfig(
+		auth.AppModuleBasic{},
+	)
+
+	kr := keyring.NewInMemory(e.Codec)
+
+	cfg := sdk.NewConfig()
+	cfg.SetCoinType(coinType)
+	cfg.SetBech32PrefixForAccount(bech32, bech32+sdk.PrefixPublic)
+
+	r, err := kr.NewAccount("tempkey", mnemonic, "", cfg.GetFullBIP44Path(), hd.Secp256k1)
+	if err != nil {
+		panic(err)
+	}
+
+	bz, err := r.GetAddress()
+	if err != nil {
+		panic(err)
+	}
+
+	bech32Addr, err := codec.NewBech32Codec(bech32).BytesToString(bz)
+	if err != nil {
+		panic(err)
+	}
+
+	return bech32Addr
 }
